+++ src/main/java/mod/gottsch/neoforge/everfurnace/core/util/CookResult.java
@@ -1,0 +1,28 @@
+/*
+ * This file is part of EverFurnace.
+ * Copyright (c) 2026 Mark Gottschling (gottsch)
+ *
+ * EverFurnace is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * EverFurnace is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with EverFurnace.  If not, see <http://www.gnu.org/licenses/lgpl>.
+ */
+package mod.gottsch.neoforge.everfurnace.core.util;
+
+/**
+ * Immutable result returned by {@code applyCookTime()}.
+ *
+ * @param itemsCooked number of items fully smelted during this catch-up pass
+ * @param xpEarned    total XP earned during this catch-up pass (may be fractional)
+ *
+ * @author by Mark Gottschling on 3/31/2026
+ */
+public record CookResult(int itemsCooked, float xpEarned) {}